requirements.txt
python-telegram-bot==21.0
openai==1.0.0
notion-client==2.2.1
python-dotenv==1.0.0